+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/balance/BalanceFragment.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.bankwallet.modules.balance2
+package io.horizontalsystems.bankwallet.modules.balance
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,7 +7,7 @@ import android.view.ViewGroup
 import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.ViewCompositionStrategy
 import io.horizontalsystems.bankwallet.core.BaseFragment
-import io.horizontalsystems.bankwallet.modules.balance2.ui.BalanceScreen
+import io.horizontalsystems.bankwallet.modules.balance.ui.BalanceScreen
 import io.horizontalsystems.core.findNavController
 
 class BalanceFragment : BaseFragment() {
