+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/balance/BalanceModule.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.bankwallet.modules.balance2
+package io.horizontalsystems.bankwallet.modules.balance
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
@@ -6,29 +6,20 @@ import io.horizontalsystems.bankwallet.core.AdapterState
 import io.horizontalsystems.bankwallet.core.App
 import io.horizontalsystems.bankwallet.core.BalanceData
 import io.horizontalsystems.bankwallet.entities.Wallet
-import io.horizontalsystems.bankwallet.modules.balance.*
 import io.horizontalsystems.marketkit.models.CoinPrice
 
 object BalanceModule {
     class AccountsFactory : ViewModelProvider.Factory {
         @Suppress("UNCHECKED_CAST")
         override fun <T : ViewModel> create(modelClass: Class<T>): T {
-//            val activeAccountService = ActiveAccountService(App.accountManager)
-
-//            val rateAppService = RateAppService(App.rateAppManager)
-
             return BalanceAccountsViewModel(App.accountManager) as T
         }
     }
 
-    class BalanceXxxFactory : ViewModelProvider.Factory {
+    class Factory : ViewModelProvider.Factory {
         @Suppress("UNCHECKED_CAST")
         override fun <T : ViewModel> create(modelClass: Class<T>): T {
-//            val activeAccountService = ActiveAccountService(App.accountManager)
-
-//            val rateAppService = RateAppService(App.rateAppManager)
-
-            val balanceService2 = BalanceService2(
+            val balanceService = BalanceService(
                 BalanceActiveWalletRepository(App.walletManager, App.accountSettingManager),
                 BalanceXRateRepository(App.currencyManager, App.marketKit),
                 BalanceAdapterRepository(App.adapterManager, BalanceCache(App.appDatabase.enabledWalletsCacheDao())),
@@ -39,11 +30,7 @@ object BalanceModule {
                 App.accountManager
             )
 
-
-            return BalanceViewModel(
-                balanceService2,
-                BalanceViewItemFactory()
-            ) as T
+            return BalanceViewModel(balanceService, BalanceViewItemFactory()) as T
         }
     }
