+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/balance/ui/BalanceItemsEmptyWatchAccount.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.bankwallet.modules.balance2.ui
+package io.horizontalsystems.bankwallet.modules.balance.ui
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
@@ -13,14 +13,11 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import androidx.navigation.NavController
 import io.horizontalsystems.bankwallet.R
-import io.horizontalsystems.bankwallet.core.slideFromRight
 import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
-import io.horizontalsystems.bankwallet.ui.compose.components.ButtonSecondaryDefault
 
 @Composable
-fun BalanceItemsEmptyRegularAccount(navController: NavController) {
+fun BalanceItemsEmptyWatchAccount() {
     Column(
         modifier = Modifier.fillMaxSize(),
         horizontalAlignment = Alignment.CenterHorizontally,
@@ -34,25 +31,19 @@ fun BalanceItemsEmptyRegularAccount(navController: NavController) {
             contentAlignment = Alignment.Center
         ) {
             Icon(
-                painter = painterResource(id = R.drawable.ic_add_to_wallet_2_48),
+                painter = painterResource(id = R.drawable.ic_emptywallet_48),
                 contentDescription = null,
                 tint = ComposeAppTheme.colors.grey
             )
         }
-        Spacer(modifier = Modifier.height(26.dp))
+        Spacer(modifier = Modifier.height(32.dp))
         Text(
             modifier = Modifier.padding(horizontal = 48.dp),
-            text = stringResource(id = R.string.Balance_NoCoinsAlert),
+            text = stringResource(id = R.string.Balance_WatchAccount_NoBalance),
             color = ComposeAppTheme.colors.grey,
             style = ComposeAppTheme.typography.subhead2,
             textAlign = TextAlign.Center
         )
-        Spacer(modifier = Modifier.height(38.dp))
-        ButtonSecondaryDefault(
-            title = stringResource(id = R.string.Balance_AddCoins),
-            onClick = {
-                navController.slideFromRight(R.id.manageWalletsFragment)
-            }
-        )
     }
+
 }
