+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/balance/ui/BalanceScreen.kt
@@ -1,10 +1,10 @@
-package io.horizontalsystems.bankwallet.modules.balance2.ui
+package io.horizontalsystems.bankwallet.modules.balance.ui
 
 import androidx.compose.runtime.Composable
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
-import io.horizontalsystems.bankwallet.modules.balance2.BalanceAccountsViewModel
-import io.horizontalsystems.bankwallet.modules.balance2.BalanceModule
+import io.horizontalsystems.bankwallet.modules.balance.BalanceAccountsViewModel
+import io.horizontalsystems.bankwallet.modules.balance.BalanceModule
 import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
 
 @Composable
