+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/main/MainViewPagerAdapter.kt
@@ -2,7 +2,7 @@ package io.horizontalsystems.bankwallet.modules.main
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import io.horizontalsystems.bankwallet.modules.balance2.BalanceFragment
+import io.horizontalsystems.bankwallet.modules.balance.BalanceFragment
 import io.horizontalsystems.bankwallet.modules.market.MarketFragment
 import io.horizontalsystems.bankwallet.modules.settings.main.MainSettingsFragment
 import io.horizontalsystems.bankwallet.modules.transactions.TransactionsFragment
